(GH-181) Move CLI options to GitReleaseManager.Core library
