+++ Source/GitReleaseManager/Options/DiscardSubOptions.cs
@@ -1,15 +1,15 @@
 //-----------------------------------------------------------------------
-// <copyright file="OpenSubOptions.cs" company="GitTools Contributors">
+// <copyright file="DiscardOptions.cs" company="GitTools Contributors">
 //     Copyright (c) 2015 - Present - GitTools Contributors
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace GitReleaseManager.Cli.Options
+namespace GitReleaseManager.Core.Options
 {
     using CommandLine;
 
-    [Verb("open", HelpText = "Opens the milestone.")]
-    public class OpenSubOptions : BaseVcsOptions
+    [Verb("discard", HelpText = "Discards a draft release.")]
+    public class DiscardSubOptions : BaseVcsOptions
     {
         [Option('m', "milestone", HelpText = "The milestone to use.", Required = true)]
         public string Milestone { get; set; }
