+++ Source/GitReleaseManager/Options/LabelSubOptions.cs
@@ -1,15 +1,15 @@
 //-----------------------------------------------------------------------
-// <copyright file="InitSubOptions.cs" company="GitTools Contributors">
+// <copyright file="LabelSubOptions.cs" company="GitTools Contributors">
 //     Copyright (c) 2015 - Present - GitTools Contributors
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace GitReleaseManager.Cli.Options
+namespace GitReleaseManager.Core.Options
 {
     using CommandLine;
 
-    [Verb("init", HelpText = "Creates a sample Yaml Configuration file in root directory")]
-    public class InitSubOptions : BaseSubOptions
+    [Verb("label", HelpText = "Deletes existing labels and replaces with set of default labels.")]
+    public class LabelSubOptions : BaseVcsOptions
     {
     }
 }
