+++ Source/GitReleaseManager/Options/OpenSubOptions.cs
@@ -1,15 +1,15 @@
 //-----------------------------------------------------------------------
-// <copyright file="CloseSubOptions.cs" company="GitTools Contributors">
+// <copyright file="OpenSubOptions.cs" company="GitTools Contributors">
 //     Copyright (c) 2015 - Present - GitTools Contributors
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace GitReleaseManager.Cli.Options
+namespace GitReleaseManager.Core.Options
 {
     using CommandLine;
 
-    [Verb("close", HelpText = "Closes the milestone.")]
-    public class CloseSubOptions : BaseVcsOptions
+    [Verb("open", HelpText = "Opens the milestone.")]
+    public class OpenSubOptions : BaseVcsOptions
     {
         [Option('m', "milestone", HelpText = "The milestone to use.", Required = true)]
         public string Milestone { get; set; }
