+++ Source/GitReleaseManager/Options/ShowConfigSubOptions.cs
@@ -1,15 +1,15 @@
 //-----------------------------------------------------------------------
-// <copyright file="LabelSubOptions.cs" company="GitTools Contributors">
+// <copyright file="ShowConfigSubOptions.cs" company="GitTools Contributors">
 //     Copyright (c) 2015 - Present - GitTools Contributors
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace GitReleaseManager.Cli.Options
+namespace GitReleaseManager.Core.Options
 {
     using CommandLine;
 
-    [Verb("label", HelpText = "Deletes existing labels and replaces with set of default labels.")]
-    public class LabelSubOptions : BaseVcsOptions
+    [Verb("showconfig", HelpText = "Shows the current configuration")]
+    public class ShowConfigSubOptions : BaseSubOptions
     {
     }
 }
